refactoring : demenTest 도메인 삭제
