+++ src/main/java/com/ssafy/soljigi/diagnosis/controller/DiagnosisController.java
@@ -5,8 +5,12 @@ import java.util.Map;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.ssafy.soljigi.diagnosis.dto.ResultSaveDTO;
 import com.ssafy.soljigi.diagnosis.service.AttentionService;
 import com.ssafy.soljigi.diagnosis.service.ExecutiveService;
 import com.ssafy.soljigi.diagnosis.service.MemoryService;
@@ -38,5 +42,17 @@ public class DiagnosisController {
 		model.addAttribute("executive", executiveService.getQuiz());
 
 		return "diagnosis/diagnosis";
+	}
+
+	@GetMapping("/diagnosis/resultPage")
+	public String getDiagnosisPage() {
+		return "diagnosis/diagnosisResult";
+	}
+
+	@ResponseBody
+	@PostMapping("diagnosis/saveResult")
+	public ResultSaveDTO save(@RequestBody ResultSaveDTO saveDTO) {
+		System.out.println(saveDTO.getOrient() + " " + saveDTO.getAttention());
+		return saveDTO;
 	}
 }
