+++ src/main/resources/static/css/diagnosis/diagnosisResult.css
