+++ src/main/resources/templates/diagnosis/diagnosisResult.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <link rel="stylesheet" type="text/css" th:href="@{/css/demenTest/demenTester.css}"/>
-    <script src="/js/demenTest/demenTester.js" defer></script>
+    <link rel="stylesheet" type="text/css" th:href="@{/css/diagnosis/diagnosisResult.css}"/>
+    <script src="/js/diagnosis/diagnosisResult.js" defer></script>
     <title>치매진단검사</title>
 </head>
 <!-- /header -->
